+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Optional iCloud Keychain sync for connection passwords
+
 ## [0.20.2] - 2026-03-18
 
 ### Fixed